user_service/requirements.txt
fastapi
pymongo
python-jose
passlib[bcrypt]
uvicorn

uvicorn
PyJWT
